Add resolved source info in ResolutionRequest status
Related to
- https://github.com/tektoncd/pipeline/issues/5529
- https://github.com/tektoncd/pipeline/pull/5397

Before:
The customized status of ResolutionRequest only contains the
resolved data.

Now:
The resolved source reference of the remote data is also added
to the ResolutionRequest.status. It is recorded in a structured way
using the standard SLSA ConfigSource struct.

Why?
Recently there is a clear requirement that the remote source
information of the remote data should be recorded in the provenance to
link back to its origin including the resolved the commit sha when users
only provide the branch/tag name for the resolver. Without this PR, the
only way to achieve this is to pass the resolved source information through
annotations, which has a couple of the drawbacks i.e. unstructured data,
hard to maintain and to make changes in future etc. That's where this PR
comes in to solve the problem.

Signed-off-by: Chuang Wang <chuangw@google.com>